tests/wpt/tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-first-valid-applies.html

<!doctype html>
<title>Multiple color-scheme meta tags - first valid applies</title>
<meta name="color-scheme">
<meta name="color-scheme" content>
<meta name="color-scheme" content="">
<meta name="color-scheme" content="light,dark">
<!-- This is first with a valid content value -->
<meta name="color-scheme" content="dark">
<meta name="color-scheme" content="light">
<link rel="help" href="https://html.spec.whatwg.org/multipage/semantics.html#meta-color-scheme">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<script>
  assert_root_color_scheme("dark", "Tree order decides which meta color-scheme applies.");
</script>
