LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-first-valid-applies-expected.txt


FAIL Tree order decides which meta color-scheme applies. assert_equals: expected (string) "dark" but got (undefined) undefined
